% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Adult_PBMC}
\alias{Adult_PBMC}
\title{Adult (PBMC) [GSE67170]}
\format{A data frame with 2 columuns, mean (mu) and variance (var), and 485574 rows for CpGs.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE67170}
}
\usage{
Adult_PBMC
}
\description{
Methylation data for GSE67170 was obtained from the genomics data repository GEO. The beta-values for 20/89 patients (disease state: control) were used to calculate CpG-specific mean and variance for 485574 CpGs.
}
\keyword{datasets}
