% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Blood_5yrOlds}
\alias{Blood_5yrOlds}
\title{Blood 5 year olds (GSE83334)}
\format{A data frame with 2 columuns, mean (mu) and variance (var), and 485512 rows for CpGs.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE83334}
}
\usage{
Blood_5yrOlds
}
\description{
Methylation data for GSE83334 was obtained from the genomics data repository GEO. The beta-values for 15/30 patients (age: 5 years) were used to calculate CpG-specific mean and variance for 485512 CpGs.
}
\keyword{datasets}
