% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Blood_adult}
\alias{Blood_adult}
\title{Blood (Adults) - Peripheral Leukocytes [GSE42861]}
\format{A data frame with 2 columuns, mean (mu) and variance (var), and 485577 rows for CpGs.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE42861}
}
\usage{
Blood_adult
}
\description{
Methylation data for GSE42861 was obtained from the genomics data repository GEO. The beta-values for 335/689 patients (subject: Normal) were used to calculate CpG-specific mean and variance for 485577 CpGs.
}
\keyword{datasets}
