% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BloodNewborns}
\alias{BloodNewborns}
\title{Blood newborns (GSE82273)}
\format{A data frame with 2 columuns, mean (mu) and variance (var), and 485577 rows for CpGs.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE82273}
}
\usage{
BloodNewborns
}
\description{
Methylation data for GSE82273 was obtained from the genomics data repository GEO. The beta-values for all 889 patients were used to calculate CpG-specific mean and variance for 485577 CpGs.
}
\keyword{datasets}
