% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Colon}
\alias{Colon}
\title{Colon (GSE77718)}
\format{A data frame with 2 columuns, mean (mu) and variance (var), and 371377 rows for CpGs.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE77718}
}
\usage{
Colon
}
\description{
Methylation data for GSE77718 was obtained from the genomics data repository GEO. The beta-values for 96/192 patients (disease state: Normal) were used to calculate CpG-specific mean and variance for 371377 CpGs.
}
\keyword{datasets}
