% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CordBlood_PBMC}
\alias{CordBlood_PBMC}
\title{Cord-blood (PBMC) [GSE110128]}
\format{A data frame with 2 columuns, mean (mu) and variance (var), and 485512 rows for CpGs.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE110128}
}
\usage{
CordBlood_PBMC
}
\description{
Methylation data for GSE110128 was obtained from the genomics data repository GEO. The beta-values for 60/233 patients (cord blood and Control) were used to calculate CpG-specific mean and variance for 485512 CpGs.
}
\keyword{datasets}
