% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CordBlood_wholeBlood}
\alias{CordBlood_wholeBlood}
\title{Cord-blood (whole blood) [GSE69176]}
\format{A data frame with 2 columuns, mean (mu) and variance (var), and 473673 rows for CpGs.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE69176}
}
\usage{
CordBlood_wholeBlood
}
\description{
Methylation data for GSE69176 was obtained from the genomics data repository GEO. The beta-values for all 152 patients were used to calculate CpG-specific mean and variance for 473673 CpGs.
}
\keyword{datasets}
