% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Liver}
\alias{Liver}
\title{Liver (GSE61258)}
\format{A data frame with 2 columuns, mean (mu) and variance (var), and 485577 rows for CpGs.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE61258}
}
\usage{
Liver
}
\description{
Methylation data for GSE61258 was obtained from the genomics data repository GEO. The beta-values for 26/79 patients (diseasestatus: Control) were used to calculate CpG-specific mean and variance for 485577 CpGs.
}
\keyword{datasets}
