% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lymphoma}
\alias{Lymphoma}
\title{Lymphoma (GSE42372)}
\format{A data frame with 2 columuns, mean (mu) and variance (var), and 485577 rows for CpGs.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE42372}
}
\usage{
Lymphoma
}
\description{
Methylation data for GSE42372 was obtained from the genomics data repository GEO. The beta-values for 20/31 patients (disease state: non-HIV lymphoma) were used to calculate CpG-specific mean and variance for 485577 CpGs.
}
\keyword{datasets}
