% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunction_v1.3.R
\name{myDensityPlots}
\alias{myDensityPlots}
\title{Density plot for simulated differences in mean methylation}
\usage{
myDensityPlots(data, detectionLimit = 0.01)
}
\arguments{
\item{data}{"deltaArray" attribute within the pwrEWAS object create by pwrEWAS}
}
\description{
myDensityPlots create a density plot of the simulated differences in mean methylation for different effect sizes
}
\examples{
myDensityPlots(data = out$deltaArray, detectionLimit = 0.01)
}
