% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunction_v1.3.R
\name{myPlotCI3D}
\alias{myPlotCI3D}
\title{Plot funtion to create a power plot}
\usage{
myPlotCI3D(data)
}
\arguments{
\item{data}{"powerArray" attribute within the pwrEWAS object create by pwrEWAS}
}
\description{
myPlotCI3D create a figure with power (with 95-percentile interval (2.5% & 97.5%)) as a funtion sample size for different effect sizes
}
\examples{
myPlotCI3D(data = out$powerArray)
}
