% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwrEWAS.shiny_v1.5.R
\name{pwrEWAS.shiny}
\alias{pwrEWAS.shiny}
\title{Shiny pwrEWAS}
\usage{
pwrEWAS.shiny()
}
\description{
pwrEWAS.shiny provides a user-friendly point-and-click interface for pwrEWAS
}
\keyword{DNAm}
\keyword{Shiny}
\keyword{microarray}
\keyword{power}
