% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunction_v1.4.R
\name{pwrEWAS_deltaDensity}
\alias{pwrEWAS_deltaDensity}
\title{Density plot for simulated differences in mean methylation}
\usage{
pwrEWAS_deltaDensity(data, detectionLimit = 0.01, sd = FALSE)
}
\arguments{
\item{data}{"deltaArray" attribute within the pwrEWAS object create by pwrEWAS}

\item{detectionLimit}{Detection limit specified in pwrEWAS.}

\item{sd}{FALSE if targetDelta was specified in pwrEWAS, and TRUE if deltaSD was specified in pwrEWAS.}
}
\description{
pwrEWAS_deltaDensity create a density plot of the simulated differences in mean methylation for different effect sizes
}
\examples{
out = pwrEWAS(minTotSampleSize = 10,
                   maxTotSampleSize = 20,
                   SampleSizeSteps = 10,
                   NcntPer = 0.5,
                   targetDelta = c(0.2, 0.5),
                   J = 1000,
                   targetDmCpGs = 10,
                   tissueType = "Adult (PBMC)",
                   detectionLimit = 0.01,
                   DMmethod = "limma",
                   FDRcritVal = 0.05,
                   core = 2,
                   sims = 30)
pwrEWAS_deltaDensity(data = out$deltaArray, detectionLimit = 0.01, sd = FALSE)
out = pwrEWAS(minTotSampleSize = 10,
                   maxTotSampleSize = 20,
                   SampleSizeSteps = 10,
                   NcntPer = 0.5,
                   deltaSD = c(0.02, 0.03),
                   J = 1000,
                   targetDmCpGs = 10,
                   tissueType = "Adult (PBMC)",
                   detectionLimit = 0.01,
                   DMmethod = "limma",
                   FDRcritVal = 0.05,
                   core = 2,
                   sims = 30)
pwrEWAS_deltaDensity(data = out$deltaArray, detectionLimit = 0.01, sd = TRUE)
}
