% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunction_v1.4.R
\name{pwrEWAS_powerPlot}
\alias{pwrEWAS_powerPlot}
\title{Plot funtion to create a power plot}
\usage{
pwrEWAS_powerPlot(data, sd = FALSE)
}
\arguments{
\item{data}{"powerArray" attribute within the pwrEWAS object create by pwrEWAS.}

\item{sd}{FALSE if targetDelta was specified in pwrEWAS, and TRUE if deltaSD was specified in pwrEWAS.}
}
\value{
pwrEWAS_powerPlot return a figure displaying power as a funtion sample size for different effect sizes
}
\description{
pwrEWAS_powerPlot create a figure with power (with 95-percentile interval (2.5% & 97.5%)) as a funtion sample size for different effect sizes
}
\examples{
outDelta = pwrEWAS(minTotSampleSize = 10,
    maxTotSampleSize = 20,
    SampleSizeSteps = 10,
    NcntPer = 0.5,
    targetDelta = c(0.2, 0.5),
    J = 1000,
    targetDmCpGs = 10,
    tissueType = "Adult (PBMC)",
    detectionLimit = 0.01,
    DMmethod = "limma",
    FDRcritVal = 0.05,
    core = 2,
    sims = 30)
pwrEWAS_powerPlot(data = outDelta$powerArray, sd = FALSE)
outSD = pwrEWAS(minTotSampleSize = 10,
    maxTotSampleSize = 20,
    SampleSizeSteps = 10,
    NcntPer = 0.5,
    deltaSD = c(0.02, 0.03),
    J = 1000,
    targetDmCpGs = 10,
    tissueType = "Adult (PBMC)",
    detectionLimit = 0.01,
    DMmethod = "limma",
    FDRcritVal = 0.05,
    core = 2,
    sims = 30)
pwrEWAS_powerPlot(data = outSD$powerArray, sd = TRUE)
}
