% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwrEWAS.shiny_v1.7.R
\name{pwrEWAS_shiny}
\alias{pwrEWAS_shiny}
\title{Shiny pwrEWAS}
\usage{
pwrEWAS_shiny()
}
\value{
pwrEWAS_shiny initializes pwrEWAS's user-interface
}
\description{
pwrEWAS_shiny provides a user-friendly point-and-click interface for pwrEWAS
}
\examples{

if(interactive()) {
    pwrEWAS_shiny()
}
}
\keyword{DNAm}
\keyword{Shiny}
\keyword{microarray}
\keyword{power}
