% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Saliva}
\alias{Saliva}
\title{Saliva (GSE92767)}
\format{A data frame with 2 columuns, mean (mu) and variance (var), and 485575 rows for CpGs.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE92767}
}
\usage{
Saliva
}
\description{
Methylation data for GSE92767 was obtained from the genomics data repository GEO. The beta-values for all 54 patients were used to calculate CpG-specific mean and variance for 485575 CpGs.
}
\keyword{datasets}
