% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sperm}
\alias{Sperm}
\title{Sperm (GSE114753)}
\format{A data frame with 2 columuns, mean (mu) and variance (var), and 485512 rows for CpGs.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE114753}
}
\usage{
Sperm
}
\description{
Methylation data for GSE114753 was obtained from the genomics data repository GEO. The beta-values for all 78/154 patients were used to calculate CpG-specific mean and variance for 485512 CpGs.
}
\keyword{datasets}
