% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunction_v1.4.R
\name{pwrEWAS_deltaDensity}
\alias{pwrEWAS_deltaDensity}
\title{Density plot for simulated differences in mean methylation}
\usage{
pwrEWAS_deltaDensity(data, detectionLimit = 0.01, sd = FALSE)
}
\arguments{
\item{data}{"deltaArray" attribute within the pwrEWAS object create by pwrEWAS}

\item{detectionLimit}{Detection limit specified in pwrEWAS.}

\item{sd}{FALSE if targetDelta was specified in pwrEWAS, and TRUE if deltaSD was specified in pwrEWAS.}
}
\description{
pwrEWAS_deltaDensity create a density plot of the simulated differences in mean methylation for different effect sizes
}
\examples{
# pwrEWAS_deltaDensity(data = outDelta$deltaArray, detectionLimit = 0.01, sd = FALSE)
# pwrEWAS_deltaDensity(data = outSD$deltaArray, detectionLimit = 0.01, sd = TRUE)
}
